Close current tab on mod+w
If mod+w is pressed, a dialog asks if changes should be saved. If the
user clicks “Ok”, changes are saved. If the user clicks “Cancel”,
Codebot should keep the tab open, but right now closing the tab and NOT
saving its content.

Some love is required to implement the “Cancel to remain open” behavior.